Fix leaderboard request
Epic now expects a content type of JSON and for us to send an empty array as the body. This is in line with their other leaderboard requests sending the user IDs from your cohort group. Since it's the global leaderboard request, it just expects nothing to be provided, but if it's not, the request breaks.